backend/api/ai_engine.py
import random
import time
import os
import hashlib
import google.generativeai as genai
from django.conf import settings
from PIL import Image
import json

class MedicalAIEngine:
    """
    A service that provides AI analysis logic.
    Integrates with Google Gemini for real analysis if key is provided.
    Falls back to deterministic simulation otherwise.
    """

    def __init__(self):
        self.model_version = "Neuro-Quant v4.2 (Gemini-Powered)"
        self.model = None

    def _get_api_key(self):
        """
        Refresh API key from DB or Settings.
        """
        try:
            from .models import SystemSettings
            from django.conf import settings
            
            settings_obj = SystemSettings.objects.first()
            if settings_obj and settings_obj.gemini_api_key:
                return settings_obj.gemini_api_key
            
            return getattr(settings, 'GEMINI_API_KEY', None)
        except Exception as e:
            print(f"API Key fetch error: {e}")
        return None

    def analyze_image(self, file_path):
        """
        Analyzes the image and returns a comprehensive medical report.
        Tries multiple Gemini models in order of capability.
        """
        api_key = self._get_api_key()

        if api_key:
            genai.configure(api_key=api_key)
            
            # List of models to try in order of preference
            models_to_try = [
                'gemini-1.5-pro',
                'gemini-1.5-flash',
                'gemini-1.0-pro-vision',
                'gemini-pro-vision'
            ]

            img = None
            try:
                # Load image (DICOM logic)
                if file_path.lower().endswith('.dcm'):
                    import pydicom
                    import numpy as np
                    ds = pydicom.dcmread(file_path)
                    pixel_array = ds.pixel_array
                    if pixel_array.max() > 0:
                        pixel_array = ((pixel_array - pixel_array.min()) / (pixel_array.max() - pixel_array.min()) * 255).astype(np.uint8)
                    img = Image.fromarray(pixel_array)
                else:
                    img = Image.open(file_path)
            except Exception as e:
                print(f"Image load error: {e}")
                img = None

            if img:
                prompt = """
                You are a sophisticated medical AI imaging assistant. Analyze this MRI/medical scan image.
                The image might be a converted DICOM frame.
                Provide a JSON response with the following structure:
                {
                    "diagnosis": "Short technical diagnosis title",
                    "accuracy": 0.95 (number between 0.90 and 0.999),
                    "workflow_speed": "45s",
                    "findings": [
                        {
                            "type": "Specific Feature (e.g., T1 Signal, Voxel Density)",
                            "status": "NORMAL/ABNORMAL/CLEAR",
                            "value": "Quantitative value (e.g., 1.2mm)",
                            "description": "Short clinical observation"
                        }
                    ],
                    "system_version": "Gemini Medical Adapter v1.0",
                    "processed_at": "current iso date"
                }
                Return ONLY the JSON. Do not include markdown formatting.
                """

                for model_name in models_to_try:
                    try:
                        print(f"Attempting analysis with model: {model_name}")
                        generative_model = genai.GenerativeModel(model_name)
                        response = generative_model.generate_content([prompt, img])
                        text = response.text.replace('```json', '').replace('```', '').strip()
                        
                        result = json.loads(text)
                        
                        # Add metadata about which model was used
                        if 'system_version' in result:
                            result['system_version'] = f"{result['system_version']} ({model_name})"
                        if 'processed_at' not in result:
                            result['processed_at'] = time.strftime("%Y-%m-%d %H:%M:%S")
                            
                        print(f"Success with {model_name}")
                        return result
                        
                    except Exception as e:
                        print(f"Failed with {model_name}: {e}")
                        continue
                
                print("All Gemini models failed. Falling back to simulation.")

        # Deterministic Fallback Simulation
        seed_val = int(hashlib.sha256(file_path.encode('utf-8')).hexdigest(), 16)
        rng = random.Random(seed_val)
        
        time.sleep(rng.uniform(2.0, 5.0))

        return {
            "diagnosis": "Neuro-Quant Automated Analysis Protocol (Simulation)",
            "accuracy": round(rng.uniform(0.97, 0.999), 4),
            "workflow_speed": f"{round(rng.uniform(30, 60), 1)}s",
            "findings": [
                {
                    "type": "Voxel Density T1",
                    "status": "NORMAL",
                    "value": "1.24 mm³",
                    "description": "Korton qalinligi standart anatomik darajada. Neyronlar zichligi optimal."
                },
                {
                    "type": "O'sma (Tumor) Detection",
                    "status": "CLEAR",
                    "value": "NEGATIVE",
                    "description": "T1-vaznli skanerlashda sezilarli anomaliyalar aniqlanmadi. Mass-effekt kuzatilmadi."
                },
                {
                    "type": "Segmentatsiya Tahlili",
                    "status": "OPTIMAL",
                    "value": "COMPLETED",
                    "description": "Loplar segmentatsiyasida anatomik dislokatsiya aniqlanmadi."
                },
                {
                    "type": "Technical Acquisition Integrity",
                    "status": "VALID",
                    "value": "SNR: 42.5dB",
                    "description": "Skanerlash sifati diagnostik analiz uchun yetarli darajada yuqori."
                }
            ],
            "system_version": self.model_version,
            "processed_at": time.strftime("%Y-%m-%d %H:%M:%S")
        }

ai_engine = MedicalAIEngine()
