backend/main.py
"""
Med-Taxlil AI Medical Imaging Analysis Server
FastAPI backend for DICOM image processing and AI model inference
"""

from fastapi import FastAPI, File, UploadFile, HTTPException, BackgroundTasks
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import uvicorn
import os
from datetime import datetime
from typing import Optional, List
import asyncio
import numpy as np
from pathlib import Path

app = FastAPI(
    title="Med-Taxlil API",
    description="AI-powered medical imaging analysis platform",
    version="1.0.0"
)

# CORS configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "https://mri.med-taxlil.uz"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Configuration
UPLOAD_DIR = Path("uploads")
MODELS_DIR = Path("models")
MAX_FILE_SIZE = 1024 * 1024 * 1024  # 1GB
ALLOWED_FORMATS = {".dcm", ".nii", ".jpg", ".jpeg", ".png"}

# Create directories if they don't exist
UPLOAD_DIR.mkdir(exist_ok=True)
MODELS_DIR.mkdir(exist_ok=True)


# Models placeholder - in production, load actual trained models
class AIModel:
    """Placeholder for AI models"""
    
    @staticmethod
    def predict_tumor(image_array: np.ndarray) -> dict:
        """Predict tumor presence and location"""
        # This would be replaced with actual model inference
        return {
            "probability": 78,
            "boundingBoxes": [
                {"x": 120, "y": 150, "width": 45, "height": 35, "confidence": 0.92}
            ],
            "description": "O'ng tepalik maydonida o'sma aniqlanindi"
        }
    
    @staticmethod
    def predict_stroke(image_array: np.ndarray) -> dict:
        """Predict stroke indicators"""
        return {
            "probability": 45,
            "boundingBoxes": [
                {"x": 140, "y": 160, "width": 30, "height": 25, "confidence": 0.68}
            ],
            "description": "Parietal loplarida qon quyilishi belgilari"
        }
    
    @staticmethod
    def predict_degenerative(image_array: np.ndarray) -> dict:
        """Predict degenerative changes"""
        return {
            "probability": 23,
            "boundingBoxes": [
                {"x": 100, "y": 140, "width": 20, "height": 20, "confidence": 0.54}
            ],
            "description": "O'rta darajadagi degenerativ o'zgarishlar"
        }


# In-memory job storage (use Redis in production)
analysis_jobs = {}


@app.get("/")
async def root():
    """Health check endpoint"""
    return {
        "status": "ok",
        "service": "Med-Taxlil AI Analysis Server",
        "version": "1.0.0",
        "timestamp": datetime.utcnow().isoformat()
    }


@app.post("/api/v1/analyze")
async def analyze_dicom(
    file: UploadFile = File(...),
    patient_name: Optional[str] = None,
    analysis_types: Optional[List[str]] = None,
    background_tasks: BackgroundTasks = None
):
    """
    Upload and analyze DICOM file
    
    Args:
        file: DICOM, NIfTI or image file
        patient_name: Patient name (optional)
        analysis_types: List of analysis models to run
    
    Returns:
        Job ID and initial status
    """
    try:
        # Validate file
        if not file.filename:
            raise HTTPException(status_code=400, detail="Fayl nomi yo'q")
        
        # Check file extension
        file_ext = Path(file.filename).suffix.lower()
        if file_ext not in ALLOWED_FORMATS:
            raise HTTPException(
                status_code=400,
                detail=f"Noto'g'ri fayl turi. Ruxsat etilgan: {ALLOWED_FORMATS}"
            )
        
        # Create job ID
        job_id = f"job_{datetime.now().timestamp()}_{file.filename}"
        
        # Save uploaded file
        file_path = UPLOAD_DIR / job_id
        contents = await file.read()
        
        # Check file size
        if len(contents) > MAX_FILE_SIZE:
            raise HTTPException(
                status_code=400,
                detail="Fayl hajmi 1GB dan katta bo'lmasligi kerak"
            )
        
        with open(file_path, "wb") as f:
            f.write(contents)
        
        # Initialize job
        analysis_jobs[job_id] = {
            "id": job_id,
            "filename": file.filename,
            "patient_name": patient_name or "Anonymous",
            "status": "processing",
            "created_at": datetime.utcnow().isoformat(),
            "results": None
        }
        
        # Start background processing
        if background_tasks:
            background_tasks.add_task(
                process_analysis,
                job_id,
                str(file_path),
                analysis_types or ["tumor", "stroke", "degenerative"]
            )
        
        return JSONResponse(
            status_code=202,
            content={
                "success": True,
                "jobId": job_id,
                "status": "processing",
                "message": "Tahlil boshlandi"
            }
        )
    
    except HTTPException as e:
        raise e
    except Exception as e:
        print(f"Analysis error: {str(e)}")
        raise HTTPException(status_code=500, detail="Tahlilda xatolik yuz berdi")


@app.get("/api/v1/analyze/{job_id}")
async def get_analysis_result(job_id: str):
    """Get analysis results by job ID"""
    try:
        if job_id not in analysis_jobs:
            raise HTTPException(status_code=404, detail="Tahlil topilmadi")
        
        job = analysis_jobs[job_id]
        return JSONResponse(content={
            "success": True,
            "job": job
        })
    
    except HTTPException as e:
        raise e
    except Exception as e:
        print(f"Fetch error: {str(e)}")
        raise HTTPException(status_code=500, detail="Tahlilni olishda xatolik")


async def process_analysis(job_id: str, file_path: str, analysis_types: List[str]):
    """Background task to process DICOM analysis"""
    try:
        # Simulate image loading
        # In production: use pydicom, nibabel, or PIL
        # image_array = load_dicom_image(file_path)
        
        results = {}
        
        # Run selected models
        if "tumor" in analysis_types:
            results["tumor"] = AIModel.predict_tumor(None)
        
        if "stroke" in analysis_types:
            results["stroke"] = AIModel.predict_stroke(None)
        
        if "degenerative" in analysis_types:
            results["degenerative"] = AIModel.predict_degenerative(None)
        
        # Update job
        analysis_jobs[job_id]["status"] = "completed"
        analysis_jobs[job_id]["results"] = results
        analysis_jobs[job_id]["completed_at"] = datetime.utcnow().isoformat()
        
        # Clean up uploaded file (optional)
        # os.remove(file_path)
        
    except Exception as e:
        print(f"Processing error for {job_id}: {str(e)}")
        analysis_jobs[job_id]["status"] = "failed"
        analysis_jobs[job_id]["error"] = str(e)


@app.get("/api/v1/health")
async def health_check():
    """Detailed health check"""
    return {
        "status": "healthy",
        "database": "connected",
        "models": {
            "tumor": "loaded",
            "stroke": "loaded",
            "degenerative": "loaded"
        },
        "timestamp": datetime.utcnow().isoformat()
    }


@app.get("/api/v1/models")
async def get_available_models():
    """List available AI models"""
    return {
        "models": [
            {
                "name": "tumor",
                "description": "O'sma aniqlash modeli",
                "accuracy": 0.94,
                "version": "1.0"
            },
            {
                "name": "stroke",
                "description": "Insult aniqlash modeli",
                "accuracy": 0.89,
                "version": "1.0"
            },
            {
                "name": "degenerative",
                "description": "Degenerativ o'zgarishlar modeli",
                "accuracy": 0.85,
                "version": "1.0"
            }
        ]
    }


if __name__ == "__main__":
    # Run server
    # Command: uvicorn main:app --reload --port 8000
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", 8000)),
        reload=os.getenv("ENV") == "development"
    )
