backend/requirements.txt
# Med-Taxlil Backend Dependencies (Django REST Framework)

# Web Framework
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
drf-spectacular>=0.26.5

# Image Processing (Relaxed for Python 3.13 compatibility)
numpy>=1.26.0
Pillow>=10.0.0
opencv-python>=4.8.1.78
scipy>=1.11.3

# Medical Image Processing
pydicom>=2.4.0
nibabel>=5.1.0
SimpleITK>=2.3.0

# Database
psycopg2-binary>=2.9.9

# Authentication & Security
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
aiofiles>=23.2.1

# AI/ML
google-generativeai>=0.3.2
kagglehub>=0.1.0
tensorflow>=2.15.0
